Refactor: PR 리뷰 반영 : 오타 수정
