+++ src/main/java/org/programmers/staybb/controller/HostController.java
@@ -1,4 +1,4 @@
-package org.programmers.staybb.Controller.User;
+package org.programmers.staybb.controller;
 
 import org.programmers.staybb.dto.user.HostResponse;
 import org.programmers.staybb.global.exception.EntityNotFoundException;
@@ -31,4 +31,5 @@ public class HostController {
         throws EntityNotFoundException {
         return ResponseEntity.ok(hostService.findHost(hostId));
     }
+
 }
