+++ src/main/java/org/programmers/staybb/controller/UserController.java
@@ -1,4 +1,4 @@
-package org.programmers.staybb.Controller.User;
+package org.programmers.staybb.controller;
 
 import javax.validation.Valid;
 import org.programmers.staybb.dto.user.UserRequest;
